changed the importer (It works now)
changed BigIntegers to long

+++ src/main/java/com/xebia/xke/benchmark/rest/FlutResource.java
@@ -7,7 +7,6 @@ import com.xebia.xke.benchmark.service.BenchmarkService;
 import javax.inject.Inject;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
-import java.math.BigInteger;
 
 @Singleton
 @Path("/")
@@ -19,7 +18,7 @@ public class FlutResource {
     @Path("get/{id}")
     @GET
     @Produces({MediaType.APPLICATION_JSON})
-    public SimpleFlut get(@PathParam("id") BigInteger id) {
+    public SimpleFlut get(@PathParam("id") Long id) {
         SimpleFlut flut = service.getFlut(id);
         if (flut == null) {
             throw new WebApplicationException(404);
@@ -30,7 +29,7 @@ public class FlutResource {
     @Path("put/{id}")
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    public void create(@PathParam("id") BigInteger id, SimpleFlut flut) {
+    public void create(@PathParam("id") Long id, SimpleFlut flut) {
         service.saveFlut(flut);
     }
